thin disk class
