+++ nirvana/models/bisym.py
@@ -38,6 +38,8 @@ from .util import cov_err
 from ..data.scatter import IntrinsicScatter
 from ..data.util import impose_positive_definite, inverse
 
+from .thindisk import ThinDisk
+
 #warnings.simplefilter('ignore', RuntimeWarning)
 
 def smoothing(array, weight=1):
@@ -536,7 +538,7 @@ def fit(plate, ifu, galmeta = None, daptype='HYB10-MILESHC-MASTARHC2', dr='MPL-1
     return sampler, args
 
 
-class BisymmetricDisk:
+class BisymmetricDisk(ThinDisk):
     r"""
     Model for a rotating thin disk with a bisymmetric flow; cf. Spekkens &
     Sellwood (2007, ApJ, 664, 204).
@@ -598,50 +600,25 @@ class BisymmetricDisk:
         # Velocity dispersion curve (can be None)
         self.dc = dc
 
-        # Number of "base" parameters
-        self.nbp = 6
-        # Total number of parameters
-        self.np = self.nbp + self.vt.np + self.v2t.np + self.v2r.np
+        # Instantiate the base class, which basically keeps all of the geometric
+        # parameters.  NOTE: The parametric curves above need to be defined
+        # first because instatiation of the base class calls reinit(), which
+        # sets default guess parameters.  TODO: Consider instantiating without
+        # guess parameters...
+        super().__init__()
+
+        # Total number of parameters:  Basic geometric parameters, the position
+        # angle of the bisymmetry, and the radials profiles of the velocity
+        # terms.
+        self.np = self.nbp + 1 + self.vt.np + self.v2t.np + self.v2r.np
         if self.dc is not None:
+            # Add in the velocity dispersion terms.
             self.np += self.dc.np
-        # Initialize the parameters (see reinit function)
-        self.par_err = None
         # Flag which parameters are freely fit
         self.free = np.ones(self.np, dtype=bool)
         self.nfree = np.sum(self.free)
         # This call to reinit adds the workspace attributes
         self.reinit()
-
-    def __repr__(self):
-        """
-        Provide the representation of the object when written to the screen.
-        """
-        # Collect the attributes relevant to construction of a model
-        attr = [n for n in ['par', 'x', 'y', 'sb', 'beam_fft'] if getattr(self, n) is not None]
-        return f'<{self.__class__.__name__}: Defined attr - {",".join(attr)}>'
-
-    def reinit(self):
-        """
-        Reinitialize the object.
-
-        This resets the model parameters to the guess parameters and erases any
-        existing data used to construct the models.  Note that, just like when
-        instantiating a new object, any calls to :func:`model` after
-        reinitialization will require at least the coordinates (``x`` and ``y``)
-        to be provided to successfully calculate the model.
-        """
-        self.par = self.guess_par()
-        self.x = None
-        self.y = None
-        self.beam_fft = None
-        self.kin = None
-        self.sb = None
-        self.vel_gpm = None
-        self.sig_gpm = None
-        self.cnvfftw = None
-        self.global_mask = 0
-        self.fit_status = None
-        self.fit_success = None
 
     def guess_par(self):
         """
@@ -654,8 +631,7 @@ class BisymmetricDisk:
         Returns:
             `numpy.ndarray`_: Vector of guess parameters
         """
-        # Return the 
-        gp = np.concatenate(([0., 0., 45., 30., 0., 0.], self.vt.guess_par(),
+        gp = np.concatenate((super().guess_par(), [0.], self.vt.guess_par(),
                              self.v2t.guess_par(), self.v2r.guess_par()))
         return gp if self.dc is None else np.append(gp, self.dc.guess_par())
 
@@ -670,15 +646,15 @@ class BisymmetricDisk:
         Returns:
             :obj:`list`: List of parameter name strings.
         """
+        base = super().par_names(short=short)
         if short:
-            base = ['x0', 'y0', 'pa', 'inc', 'vsys', 'pab']
+            base += ['pab']
             vt = [f'vt_{p}' for p in self.vt.par_names(short=True)]
             v2t = [f'v2t_{p}' for p in self.v2t.par_names(short=True)]
             v2r = [f'v2r_{p}' for p in self.v2r.par_names(short=True)]
             dc = [] if self.dc is None else [f's_{p}' for p in self.dc.par_names(short=True)]
         else:
-            base = ['X center', 'Y center', 'Position Angle', 'Inclination', 'Systemic Velocity',
-                    'Bisymmetry PA']
+            base += ['Bisymmetry PA']
             vt = [f'VT: {p}' for p in self.vt.par_names()]
             v2t = [f'V2T: {p}' for p in self.v2t.par_names()]
             v2r = [f'V2R: {p}' for p in self.v2r.par_names()]
@@ -687,36 +663,18 @@ class BisymmetricDisk:
 
     # These "private" functions yield slices of the parameter vector for the
     # desired set of parameters
-    def _base_slice(self):
-        return slice(self.nbp)
     def _vt_slice(self):
-        s = self.nbp
+        s = self.nbp + 1
         return slice(s, s + self.vt.np)
     def _v2t_slice(self):
-        s = self.nbp + self.vt.np
+        s = self.nbp + 1 + self.vt.np
         return slice(s, s + self.v2t.np)
     def _v2r_slice(self):
-        s = self.nbp + self.vt.np + self.v2t.np
+        s = self.nbp + 1 + self.vt.np + self.v2t.np
         return slice(s, s + self.v2r.np)
     def _dc_slice(self):
-        s = self.nbp + self.vt.np + self.v2t.np + self.v2r.np
+        s = self.nbp + 1 + self.vt.np + self.v2t.np + self.v2r.np
         return slice(s, s + self.dc.np)
-
-    def base_par(self, err=False):
-        """
-        Return the base (largely geometric) parameters. Returns None if
-        parameters are not defined yet.
-
-        Args:
-            err (:obj:`bool`, optional):
-                Return the parameter errors instead of the parameter values.
-
-        Returns:
-            `numpy.ndarray`_: Vector with parameters or parameter errors for the
-            "base" parameters.
-        """
-        p = self.par_err if err else self.par
-        return None if p is None else p[self._base_slice()]
 
     def vt_par(self, err=False):
         """
@@ -782,7 +740,7 @@ class BisymmetricDisk:
         p = self.par_err if err else self.par
         return None if p is None or self.dc is None else p[self._dc_slice()]
 
-    def par_bounds(self, base_lb=None, base_ub=None):
+    def par_bounds(self, base_lb=None, base_ub=None, pab_lb=None, pab_ub=None):
         """
         Return the lower and upper boundaries on the model parameters.
 
@@ -808,201 +766,13 @@ class BisymmetricDisk:
             :obj:`tuple`: A two-tuple providing, respectively, the lower and
             upper boundaries for all model parameters.
         """
-        if base_lb is not None and len(base_lb) != self.nbp:
-            raise ValueError('Incorrect number of lower bounds for the base '
-                             f'parameters; found {len(base_lb)}, expected {self.nbp}.')
-        if base_ub is not None and len(base_ub) != self.nbp:
-            raise ValueError('Incorrect number of upper bounds for the base '
-                             f'parameters; found {len(base_ub)}, expected {self.nbp}.')
-
-        if (base_lb is None or base_ub is None) and (self.x is None or self.y is None):
-            raise ValueError('Cannot define limits on center.  Provide base_lb,base_ub or set '
-                             'the evaluation grid coordinates (attributes x and y).')
-
-        if base_lb is None:
-            minx = np.amin(self.x)
-            miny = np.amin(self.y)
-            base_lb = np.array([minx, miny, -350., 1., -300., -100.])
-        if base_ub is None:
-            maxx = np.amax(self.x)
-            maxy = np.amax(self.y)
-            base_ub = np.array([maxx, maxy, 350., 89., 300., 100.])
-        # Minimum and maximum allowed values
-        lb = np.concatenate((base_lb, self.vt.lb, self.v2t.lb, self.v2r.lb))
-        ub = np.concatenate((base_ub, self.vt.ub, self.v2t.ub, self.v2r.ub))
+        _base_lb, _base_ub = super().par_bounds(base_lb=base_lb, base_ub=base_ub)
+        _base_lb = np.append(_base_lb, [-100. if pab_lb is None else pab_lb])
+        _base_ub = np.append(_base_ub, [100. if pab_ub is None else pab_ub])
+        lb = np.concatenate((_base_lb, self.vt.lb, self.v2t.lb, self.v2r.lb))
+        ub = np.concatenate((_base_ub, self.vt.ub, self.v2t.ub, self.v2r.ub))
         return (lb, ub) if self.dc is None \
                     else (np.append(lb, self.dc.lb), np.append(ub, self.dc.ub))
-
-    def _set_par(self, par):
-        """
-        Set the full parameter vector, accounting for any fixed parameters.
-
-        Args:
-            par (`numpy.ndarray`_, optional):
-                The list of parameters to use. Length should be either
-                :attr:`np` or :attr:`nfree`. If the latter, the values of the
-                fixed parameters in :attr:`par` are used.
-        """
-        if par.ndim != 1:
-            raise ValueError('Parameter array must be a 1D vector.')
-        if par.size == self.np:
-            self.par = par.copy()
-            return
-        if par.size != self.nfree:
-            raise ValueError('Must provide {0} or {1} parameters.'.format(self.np, self.nfree))
-        self.par[self.free] = par.copy()
-
-    def _init_coo(self, x, y):
-        """
-        Initialize the coordinate arrays.
-
-        .. warning::
-            
-            Input coordinate data types are all converted to `numpy.float64`_.
-            This is always true, even though this is only needed when using
-            :class:`~nirvana.models.beam.ConvolveFFTW`.
-
-        Args:
-            x (`numpy.ndarray`_):
-                The 2D x-coordinates at which to evaluate the model.  If not
-                None, replace the existing :attr:`x` with this array.
-            y (`numpy.ndarray`_):
-                The 2D y-coordinates at which to evaluate the model.  If not
-                None, replace the existing :attr:`y` with this array.
-
-        Raises:
-            ValueError:
-                Raised if the shapes of :attr:`x` and :attr:`y` are not the same.
-        """
-        if x is None and y is None:
-            # Nothing to do
-            return
-
-        # Define it and check it
-        if x is not None:
-            self.x = x.astype(float)
-        if y is not None:
-            self.y = y.astype(float)
-        if self.x.shape != self.y.shape:
-            raise ValueError('Input coordinates must have the same shape.')
-
-    def _init_sb(self, sb):
-        """
-        Initialize the surface brightness array.
-
-        .. warning::
-            
-            Input surface-brightness data types are all converted to
-            `numpy.float64`_.  This is always true, even though this is only
-            needed when using :class:`~nirvana.models.beam.ConvolveFFTW`.
-
-        Args:
-            sb (`numpy.ndarray`_):
-                2D array with the surface brightness of the object.  If not
-                None, replace the existing :attr:`sb` with this array.
-
-        Raises:
-            ValueError:
-                Raised if the shapes of :attr:`sb` and :attr:`x` are not the same.
-        """
-        if sb is None:
-            # Nothing to do
-            return
-
-        # Check it makes sense to define the surface brightness
-        if self.x is None:
-            raise ValueError('Input coordinates must be instantiated first!')
-
-        # Define it and check it
-        self.sb = sb.astype(float)
-        if self.sb.shape != self.x.shape:
-            raise ValueError('Input coordinates must have the same shape.')
-
-    def _init_beam(self, beam, is_fft, cnvfftw):
-        """
-        Initialize the beam-smearing kernel and the convolution method.
-
-        Args:
-            beam (`numpy.ndarray`_):
-                The 2D rendering of the beam-smearing kernel, or its Fast
-                Fourier Transform (FFT).  If not None, replace existing
-                :attr:`beam_fft` with this array (or its FFT, depending on the
-                provided ``is_fft``).
-            is_fft (:obj:`bool`):
-                The provided ``beam`` object is already the FFT of the
-                beam-smearing kernel.
-            cnvfftw (:class:`~nirvana.models.beam.ConvolveFFTW`):
-                An object that expedites the convolutions using FFTW/pyFFTW.  If
-                provided, the shape *must* match :attr:``beam_fft`` (after this
-                is potentially updated by the provided ``beam``).  If None, a
-                new :class:`~nirvana.models.beam.ConvolveFFTW` instance is
-                constructed to perform the convolutions.  If the class cannot be
-                constructed because the user doesn't have pyfftw installed, then
-                the convolutions fall back to the numpy routines.
-        """
-        if beam is None:
-            # Nothing to do
-            return
-
-        # Check it makes sense to define the beam
-        if self.x is None:
-            raise ValueError('Input coordinates must be instantiated first!')
-        if self.x.ndim != 2:
-            raise ValueError('To perform convolution, must provide 2d coordinate arrays.')
-
-        # Assign the beam and check it
-        self.beam_fft = beam if is_fft else np.fft.fftn(np.fft.ifftshift(beam))
-        if self.beam_fft.shape != self.x.shape:
-            raise ValueError('Currently, convolution requires the beam map to have the same '
-                                'shape as the coordinate maps.')
-
-        # Convolutions will be performed, try to setup the ConvolveFFTW
-        # object (self.cnvfftw).
-        if cnvfftw is None:
-            if self.cnvfftw is not None and self.cnvfftw.shape == self.beam_fft.shape:
-                # ConvolveFFTW is ready to go
-                return
-
-            try:
-                self.cnvfftw = ConvolveFFTW(self.beam_fft.shape)
-            except:
-                warnings.warn('Could not instantiate ConvolveFFTW; proceeding with numpy '
-                              'FFT/convolution routines.')
-                self.cnvfftw = None
-        else:
-            # A cnvfftw was provided, check it
-            if not isinstance(cnvfftw, ConvolveFFTW):
-                raise TypeError('Provided cnvfftw must be a ConvolveFFTW instance.')
-            if cnvfftw.shape != self.beam_fft.shape:
-                raise ValueError('cnvfftw shape does not match beam shape.')
-            self.cnvfftw = cnvfftw
-
-    def _init_par(self, p0, fix):
-        """
-        Initialize the relevant parameter vectors that tracks the full set of
-        model parameters and which of those are freely fit by the model.
-
-        Args:
-            p0 (`numpy.ndarray`_):
-                The initial parameters for the model.  Can be None.  Length must
-                be :attr:`np`, if not None.
-            fix (`numpy.ndarray`_):
-                A boolean array selecting the parameters that should be fixed
-                during the model fit.  Can be None.  Length must be :attr:`np`,
-                if not None.
-        """
-        if p0 is None:
-            p0 = self.guess_par()
-        _p0 = np.atleast_1d(p0)
-        if _p0.size != self.np:
-            raise ValueError('Incorrect number of model parameters.')
-        self.par = _p0.copy()
-        self.par_err = None
-        _free = np.ones(self.np, dtype=bool) if fix is None else np.logical_not(fix)
-        if _free.size != self.np:
-            raise ValueError('Incorrect number of model parameter fitting flags.')
-        self.free = _free.copy()
-        self.nfree = np.sum(self.free)
 
     def in_disk_bar_angle(self, par=None):
         # Reset the parameter values
@@ -1074,21 +844,8 @@ class BisymmetricDisk:
             `numpy.ndarray`_, :obj:`tuple`: The velocity field model, and the
             velocity dispersion field model, if the latter is included
         """
-        # Initialize the coordinates (this does nothing if both x and y are None)
-        self._init_coo(x, y)
-        # Initialize the surface brightness (this does nothing if sb is None)
-        self._init_sb(sb)
-        # Initialize the convolution kernel (this does nothing if beam is None)
-        self._init_beam(beam, is_fft, cnvfftw)
-        if self.beam_fft is not None and not ignore_beam:
-            # Initialize the surface brightness, only if it would be used
-            self._init_sb(sb)
-        # Check that the model can be calculated
-        if self.x is None or self.y is None:
-            raise ValueError('No coordinate grid defined.')
-        # Reset the parameter values
-        if par is not None:
-            self._set_par(par)
+        # Initialize the thin disk model
+        self._init_model(par, x, y, sb, beam, is_fft, cnvfftw, ignore_beam)
 
         # Get the coordinate data
         #   - Convert the relevant angles to radians
@@ -1124,22 +881,65 @@ class BisymmetricDisk:
     def deriv_model(self, par=None, x=None, y=None, sb=None, beam=None, is_fft=False, cnvfftw=None,
                     ignore_beam=False):
         """
+        Evaluate the derivative of the model w.r.t all input parameters.
+
+        Note that arguments passed to this function overwrite any existing
+        attributes of the object, and subsequent calls to this function will
+        continue to use existing attributes, unless they are overwritten.  For
+        example, if ``beam`` is provided here, it overwrites any existing
+        :attr:`beam_fft` and any subsequent calls to ``model`` **that do not
+        provide a new** ``beam`` will use the existing :attr:`beam_fft`.  To
+        remove all internal attributes to get a "clean" instantiation, either
+        define a new :class:`AxisymmetricDisk` instance or use :func:`reinit`.
+
+        .. warning::
+            
+            Input coordinates and surface-brightness data types are all
+            converted to `numpy.float64`_.  This is always true, even though
+            this is only needed when using
+            :class:`~nirvana.models.beam.ConvolveFFTW`.
+
+        Args:
+            par (`numpy.ndarray`_, optional):
+                The list of parameters to use. If None, the internal
+                :attr:`par` is used. Length should be either :attr:`np` or
+                :attr:`nfree`. If the latter, the values of the fixed
+                parameters in :attr:`par` are used.
+            x (`numpy.ndarray`_, optional):
+                The 2D x-coordinates at which to evaluate the model. If not
+                provided, the internal :attr:`x` is used.
+            y (`numpy.ndarray`_, optional):
+                The 2D y-coordinates at which to evaluate the model. If not
+                provided, the internal :attr:`y` is used.
+            sb (`numpy.ndarray`_, optional):
+                2D array with the surface brightness of the object. This is used
+                to weight the convolution of the kinematic fields according to
+                the luminosity distribution of the object.  Must have the same
+                shape as ``x``. If None, the convolution is unweighted.  If a
+                convolution is not performed (either ``beam`` or
+                :attr:`beam_fft` are not available, or ``ignore_beam`` is True),
+                this array is ignored.
+            beam (`numpy.ndarray`_, optional):
+                The 2D rendering of the beam-smearing kernel, or its Fast
+                Fourier Transform (FFT). If not provided, the internal
+                :attr:`beam_fft` is used.
+            is_fft (:obj:`bool`, optional):
+                The provided ``beam`` object is already the FFT of the
+                beam-smearing kernel.  Ignored if ``beam`` is not provided.
+            cnvfftw (:class:`~nirvana.models.beam.ConvolveFFTW`, optional):
+                An object that expedites the convolutions using
+                FFTW/pyFFTW. If None, the convolution is done using numpy
+                FFT routines.
+            ignore_beam (:obj:`bool`, optional):
+                Ignore the beam-smearing when constructing the model. I.e.,
+                construct the *intrinsic* model.
+
+        Returns:
+            `numpy.ndarray`_, :obj:`tuple`: The velocity field model, and the
+            velocity dispersion field model, if the latter is included
         """
-        # Initialize the coordinates (this does nothing if both x and y are None)
-        self._init_coo(x, y)
-        # Initialize the surface brightness (this does nothing if sb is None)
-        self._init_sb(sb)
-        # Initialize the convolution kernel (this does nothing if beam is None)
-        self._init_beam(beam, is_fft, cnvfftw)
-        if self.beam_fft is not None and not ignore_beam:
-            # Initialize the surface brightness, only if it would be used
-            self._init_sb(sb)
-        # Check that the model can be calculated
-        if self.x is None or self.y is None:
-            raise ValueError('No coordinate grid defined.')
-        # Reset the parameter values
-        if par is not None:
-            self._set_par(par)
+        # Initialize the thin disk model
+        self._init_model(par, x, y, sb, beam, is_fft, cnvfftw, ignore_beam)
 
         # Initialize the derivative arrays needed for the coordinate calculation
         dx = np.zeros(self.x.shape+(self.np,), dtype=float)
@@ -1233,425 +1033,6 @@ class BisymmetricDisk:
                                              sig=sig, dsig=dsig, cnvfftw=self.cnvfftw)
         return v, sig, dv, dsig
 
-
-
-#    def mock_observation(self, par, kin=None, x=None, y=None, sb=None, binid=None,
-#                         vel_ivar=None, vel_covar=None, vel_mask=None, sig_ivar=None,
-#                         sig_covar=None, sig_mask=None, sig_corr=None, beam=None, is_fft=False,
-#                         cnvfftw=None, ignore_beam=False, add_err=False, positive_definite=False,
-#                         rng=None):
-
-    def fisher_matrix(self, par, kin, fix=None, sb_wgt=False, scatter=None,
-                      assume_posdef_covar=False, ignore_covar=True, cnvfftw=None,
-                      inverse=False):
-        """
-        """
-        self._fit_prep(kin, par, fix, scatter, sb_wgt, assume_posdef_covar, ignore_covar,
-                       cnvfftw)
-        jac = self._get_jac()(self.par[self.free])
-        return cov_err(jac) if inverse else np.dot(jac.T,jac)
-
-    # This slew of "private" functions consolidate the velocity residual and
-    # chi-square calculations
-    def _v_resid(self, vel):
-        return self.kin.vel[self.vel_gpm] - vel[self.vel_gpm]
-    def _deriv_v_resid(self, dvel):
-        return -dvel[np.ix_(self.vel_gpm, self.free)]
-    def _v_chisqr(self, vel):
-        return self._v_resid(vel) / self._v_err[self.vel_gpm]
-    def _deriv_v_chisqr(self, dvel):
-        return self._deriv_v_resid(dvel) / self._v_err[self.vel_gpm, None]
-    def _v_chisqr_covar(self, vel):
-        return np.dot(self._v_resid(vel), self._v_ucov)
-    def _deriv_v_chisqr_covar(self, dvel):
-        return np.dot(self._deriv_v_resid(dvel).T, self._v_ucov).T
-
-    # This slew of "private" functions consolidate the velocity dispersion
-    # residual and chi-square calculations
-    def _s_resid(self, sig):
-        return self.kin.sig_phys2[self.sig_gpm] - sig[self.sig_gpm]**2
-    def _deriv_s_resid(self, sig, dsig):
-        return -2 * sig[self.sig_gpm,None] * dsig[np.ix_(self.sig_gpm, self.free)]
-    def _s_chisqr(self, sig):
-        return self._s_resid(sig) / self._s_err[self.sig_gpm]
-    def _deriv_s_chisqr(self, sig, dsig):
-        return self._deriv_s_resid(sig, dsig) / self._s_err[self.sig_gpm, None]
-    def _s_chisqr_covar(self, sig):
-        return np.dot(self._s_resid(sig), self._s_ucov)
-    def _deriv_s_chisqr_covar(self, sig, dsig):
-        return np.dot(self._deriv_s_resid(sig, dsig).T, self._s_ucov).T
-
-    def _binned_model(self, par):
-        """
-        Compute the binned model data.
-
-        Args:
-            par (`numpy.ndarray`_, optional):
-                The list of parameters to use. Length should be either
-                :attr:`np` or :attr:`nfree`. If the latter, the values of the
-                fixed parameters in :attr:`par` are used.
-
-        Returns:
-
-            :obj:`tuple`: Model velocity and velocity dispersion data.  The
-            latter is None if the model has no dispersion parameterization.
-        """
-        self._set_par(par)
-        if self.dc is None:
-            vel = self.model()
-            sig = None
-        else:
-            vel, sig = self.model()
-        return self.kin.bin_moments(self.sb, vel, sig)[1:]
-
-    def _deriv_binned_model(self, par):
-        """
-        Compute the binned model data and its derivatives.
-
-        Args:
-            par (`numpy.ndarray`_, optional):
-                The list of parameters to use. Length should be either
-                :attr:`np` or :attr:`nfree`. If the latter, the values of the
-                fixed parameters in :attr:`par` are used.
-
-        Returns:
-            :obj:`tuple`: Model velocity data, velocity dispersion data,
-            velocity derivative, and velocity dispersion derivative.  The
-            velocity dispersion components (2nd and 4th objects) are None if the
-            model has no dispersion parameterization.
-        """
-        self._set_par(par)
-        if self.dc is None:
-            vel, dvel = self.deriv_model()
-            sig, dsig = None, None
-        else:
-            vel, sig, dvel, dsig = self.deriv_model()
-        _, vel, sig, _, dvel, dsig \
-                = self.kin.deriv_bin_moments(self.sb, vel, sig, None, dvel, dsig)
-        return vel, sig, dvel, dsig
-
-    def _resid(self, par, sep=False):
-        """
-        Calculate the residuals between the data and the current model.
-
-        Args:
-            par (`numpy.ndarray`_, optional):
-                The list of parameters to use. Length should be either
-                :attr:`np` or :attr:`nfree`. If the latter, the values of the
-                fixed parameters in :attr:`par` are used.
-            sep (:obj:`bool`, optional):
-                Return separate vectors for the velocity and velocity
-                dispersion residuals, instead of appending them.
-
-        Returns:
-            `numpy.ndarray`_: Difference between the data and the model for
-            all measurements.
-        """
-        vel, sig = self._binned_model(par)
-        vfom = self._v_resid(vel)
-        sfom = numpy.array([]) if self.dc is None else self._s_resid(sig)
-        return (vfom, sfom) if sep else np.append(vfom, sfom)
-
-    def _deriv_resid(self, par, sep=False):
-        """
-        Calculate the derivative of the fit residuals w.r.t. all the *free*
-        model parameters.
-
-        Args:
-            par (`numpy.ndarray`_, optional):
-                The list of parameters to use. Length should be either
-                :attr:`np` or :attr:`nfree`. If the latter, the values of the
-                fixed parameters in :attr:`par` are used.
-            sep (:obj:`bool`, optional):
-                Return separate vectors for the velocity and velocity dispersion
-                residual derivatives, instead of appending them.
-
-        Returns:
-            :obj:`tuple`, `numpy.ndarray`_: Derivatives in the difference
-            between the data and the model for all measurements, either returned
-            as a single array for all data or as separate arrays for the
-            velocity and velocity dispersion data (based on ``sep``).
-        """
-        vel, sig, dvel, dsig = self._deriv_binned_model(par)
-        if self.dc is None:
-            return (self._deriv_v_resid(dvel), numpy.array([])) \
-                        if sep else self._deriv_v_resid(dvel)
-        resid = (self._deriv_v_resid(vel), self._deriv_s_resid(sig, dsig))
-        return resid if sep else np.vstack(resid)
-
-    def _chisqr(self, par, sep=False):
-        """
-        Calculate the error-normalized residual (close to the signed
-        chi-square metric) between the data and the current model.
-
-        Args:
-            par (`numpy.ndarray`_, optional):
-                The list of parameters to use. Length should be either
-                :attr:`np` or :attr:`nfree`. If the latter, the values of the
-                fixed parameters in :attr:`par` are used.
-            sep (:obj:`bool`, optional):
-                Return separate vectors for the velocity and velocity
-                dispersion residuals, instead of appending them.
-
-        Returns:
-            `numpy.ndarray`_: Difference between the data and the model for
-            all measurements, normalized by their errors.
-        """
-        vel, sig = self._binned_model(par)
-        if self.has_covar:
-            vfom = self._v_chisqr_covar(vel)
-            sfom = np.array([]) if self.dc is None else self._s_chisqr_covar(sig)
-        else:
-            vfom = self._v_chisqr(vel)
-            sfom = np.array([]) if self.dc is None else self._s_chisqr(sig)
-        return (vfom, sfom) if sep else np.append(vfom, sfom)
-
-    def _deriv_chisqr(self, par, sep=False):
-        """
-        Calculate the derivatives of the error-normalized residuals (close to
-        the signed chi-square metric) w.r.t. the *free* model parameters.
-
-        Args:
-            par (`numpy.ndarray`_, optional):
-                The list of parameters to use. Length should be either
-                :attr:`np` or :attr:`nfree`. If the latter, the values of the
-                fixed parameters in :attr:`par` are used.
-            sep (:obj:`bool`, optional):
-                Return separate vectors for the velocity and velocity
-                dispersion residuals, instead of appending them.
-
-        Returns:
-            :obj:`tuple`, `numpy.ndarray`_: Derivatives of the error-normalized
-            difference between the data and the model for all measurements
-            w.r.t. the *free* model parameters, either returned as a single
-            array for all data or as separate arrays for the velocity and
-            velocity dispersion data (based on ``sep``).
-        """
-        vel, sig, dvel, dsig = self._deriv_binned_model(par)
-        vf = self._deriv_v_chisqr_covar if self.has_covar else self._deriv_v_chisqr
-        if self.dc is None:
-            return (vf(dvel), numpy.array([])) if sep else vf(dvel)
-
-        sf = self._deriv_s_chisqr_covar if self.has_covar else self._deriv_s_chisqr
-        dchisqr = (vf(dvel), sf(sig, dsig))
-        if not np.all(np.isfinite(dchisqr[0])) or not np.all(np.isfinite(dchisqr[1])):
-            raise ValueError('Error in derivative computation.')
-        return dchisqr if sep else np.vstack(dchisqr)
-
-    def _fit_prep(self, kin, p0, fix, scatter, sb_wgt, assume_posdef_covar, ignore_covar, cnvfftw):
-        """
-        Prepare the object for fitting the provided kinematic data.
-
-        Args:
-            kin (:class:`~nirvana.data.kinematics.Kinematics`):
-                The object providing the kinematic data to be fit.
-            p0 (`numpy.ndarray`_):
-                The initial parameters for the model. Length must be
-                :attr:`np`.
-            fix (`numpy.ndarray`_):
-                A boolean array selecting the parameters that should be fixed
-                during the model fit.
-            scatter (:obj:`float`, optional):
-                Introduce a fixed intrinsic-scatter term into the model. This
-                single value is added in quadrature to all measurement errors
-                in the calculation of the merit function. If no errors are
-                available, this has the effect of renormalizing the
-                unweighted merit function by 1/scatter.
-            sb_wgt (:obj:`bool`):
-                Flag to use the surface-brightness data provided by ``kin``
-                to weight the model when applying the beam-smearing.
-            assume_posdef_covar (:obj:`bool`, optional):
-                If the :class:`~nirvana.data.kinematics.Kinematics` includes
-                covariance matrices, this forces the code to proceed assuming
-                the matrices are positive definite.
-            ignore_covar (:obj:`bool`, optional):
-                If the :class:`~nirvana.data.kinematics.Kinematics` includes
-                covariance matrices, ignore them and just use the inverse
-                variance.
-        """
-        # Initialize the fit parameters
-        self._init_par(p0, fix)
-        # Initialize the data to fit
-        self.kin = kin
-        self._init_coo(self.kin.grid_x, self.kin.grid_y)
-        self._init_sb(self.kin.grid_sb if sb_wgt else None)
-        self.vel_gpm = np.logical_not(self.kin.vel_mask)
-        self.sig_gpm = None if self.dc is None else np.logical_not(self.kin.sig_mask)
-        # Initialize the beam kernel
-        self._init_beam(self.kin.beam_fft, True, cnvfftw)
-
-        # Determine which errors were provided
-        self.has_err = self.kin.vel_ivar is not None if self.dc is None \
-                        else self.kin.vel_ivar is not None and self.kin.sig_ivar is not None
-        if not self.has_err and (self.kin.vel_err is not None or self.kin.sig_err is not None):
-            warnings.warn('Some errors being ignored if both velocity and velocity dispersion '
-                          'errors are not provided.')
-        self.has_covar = self.kin.vel_covar is not None if self.dc is None \
-                            else self.kin.vel_covar is not None and self.kin.sig_covar is not None
-        if not self.has_covar \
-                and (self.kin.vel_covar is not None or self.kin.sig_covar is not None):
-            warnings.warn('Some covariance matrices being ignored if both velocity and velocity '
-                          'dispersion covariances are not provided.')
-        if ignore_covar:
-            # Force ignoring the covariance
-            # TODO: This requires that, e.g., kin.vel_ivar also be defined...
-            self.has_covar = False
-
-        # Check the intrinsic scatter input
-        self.scatter = None
-        if scatter is not None:
-            self.scatter = np.atleast_1d(scatter)
-            if self.scatter.size > 2:
-                raise ValueError('Should provide, at most, one scatter term for each kinematic '
-                                 'moment being fit.')
-            if self.dc is not None and self.scatter.size == 1:
-                warnings.warn('Using single scatter term for both velocity and velocity '
-                              'dispersion.')
-                self.scatter = np.array([scatter, scatter])
-
-        # Set the internal error attributes
-        if self.has_err:
-            self._v_err = np.sqrt(inverse(self.kin.vel_ivar))
-            self._s_err = None if self.dc is None \
-                                else np.sqrt(inverse(self.kin.sig_phys2_ivar))
-            if self.scatter is not None:
-                self._v_err = np.sqrt(self._v_err**2 + self.scatter[0]**2)
-                if self.dc is not None:
-                    self._s_err = np.sqrt(self._s_err**2 + self.scatter[1]**2)
-        elif not self.has_err and not self.has_covar and self.scatter is not None:
-            self.has_err = True
-            self._v_err = np.full(self.kin.vel.shape, self.scatter[0], dtype=float)
-            self._s_err = None if self.dc is None \
-                                else np.full(self.kin.sig.shape, self.scatter[1], dtype=float)
-        else:
-            self._v_err = None
-            self._s_err = None
-
-        # Set the internal covariance attributes
-        if self.has_covar:
-            # Construct the matrices used to calculate the merit function in
-            # the presence of covariance.
-            vel_pd_covar = self.kin.vel_covar[np.ix_(self.vel_gpm,self.vel_gpm)]
-            sig_pd_covar = None if self.dc is None \
-                            else self.kin.sig_phys2_covar[np.ix_(self.sig_gpm,self.sig_gpm)]
-            if not assume_posdef_covar:
-                # Force the matrices to be positive definite
-                print('Forcing vel covar to be pos-def')
-                vel_pd_covar = impose_positive_definite(vel_pd_covar)
-                print('Forcing sig covar to be pos-def')
-                sig_pd_covar = None if self.dc is None else impose_positive_definite(sig_pd_covar)
-
-            if self.scatter is not None:
-                # A diagonal matrix with only positive values is, by definition,
-                # positive definite; and the sum of two positive-definite
-                # matrices is also positive definite.
-                vel_pd_covar += np.diag(np.full(vel_pd_covar.shape[0], self.scatter[0]**2,
-                                                dtype=float))
-                if self.dc is not None:
-                    sig_pd_covar += np.diag(np.full(sig_pd_covar.shape[0], self.scatter[1]**2,
-                                                    dtype=float))
-
-            self._v_ucov = cinv(vel_pd_covar, upper=True)
-            self._s_ucov = None if sig_pd_covar is None else cinv(sig_pd_covar, upper=True)
-        else:
-            self._v_ucov = None
-            self._s_ucov = None
-
-    def _get_fom(self):
-        """
-        Return the figure-of-merit function to use given the availability of
-        errors.
-        """
-        return self._chisqr if self.has_err or self.has_covar else self._resid
-
-    def _get_jac(self):
-        """
-        Return the Jacobian function to use given the availability of errors.
-        """
-        return self._deriv_chisqr if self.has_err or self.has_covar else self._deriv_resid
-
-    def lsq_fit(self, kin, sb_wgt=False, p0=None, fix=None, lb=None, ub=None, scatter=None,
-                verbose=0, assume_posdef_covar=False, ignore_covar=True, cnvfftw=None,
-                analytic_jac=True, maxiter=5):
-        """
-        """
-        if maxiter is None:
-            raise ValueError('Maximum number of iterations cannot be None.')
-
-        # Prepare to fit the data.
-        self._fit_prep(kin, p0, fix, scatter, sb_wgt, assume_posdef_covar, ignore_covar,
-                       cnvfftw)
-        
-        # Get the method used to generate the figure-of-merit and the Jacobian
-        # matrix.
-        fom = self._get_fom()
-        # If the analytic Jacobian matrix is not used, the derivative of the
-        # merit function wrt each parameter is determined by a 1% change in each
-        # parameter.
-        jac_kwargs = {'jac': self._get_jac()} if analytic_jac \
-                        else {'diff_step': np.full(self.np, 0.01, dtype=float)[self.free]}
-
-        # Parameter boundaries
-        _lb, _ub = self.par_bounds()
-        if lb is None:
-            lb = _lb
-        if ub is None:
-            ub = _ub
-        if len(lb) != self.np or len(ub) != self.np:
-            raise ValueError('Length of one or both of the bound vectors is incorrect.')
-
-        # Set the random number generator with a fixed seed so that the result
-        # is deterministic.
-        rng = np.random.default_rng(seed=909)
-        _p0 = self.par[self.free]
-        p = _p0.copy()
-        pe = None
-        niter = 0
-        while niter < maxiter:
-            # Run the optimization
-            result = optimize.least_squares(fom, p, # method='lm', #xtol=None,
-                                            x_scale='jac', method='trf', xtol=1e-12,
-                                            bounds=(lb[self.free], ub[self.free]), 
-                                            verbose=max(verbose,0), **jac_kwargs)
-            try:
-                pe = np.sqrt(np.diag(cov_err(result.jac)))
-            except:
-                warnings.warn('Unable to compute parameter errors from precision matrix.')
-                pe = None
-
-            # The fit should change the input parameters.
-            if np.all(np.absolute(p-result.x) > 1e-3):
-                break
-
-            # If it doesn't, something likely went wrong with the fit.  Perturb
-            # the input guesses a bit and retry.
-            p = _p0 + rng.normal(size=self.nfree)*(pe if pe is not None else 0.1*p0)
-            p = np.clip(p, lb[self.free], ub[self.free])
-            niter += 1
-
-        # TODO: Add something to the fit status/success flags that tests if
-        # niter == maxiter and/or if the input parameters are identical to the
-        # final best-fit prameters?  Note that the input parameters, p0, may not
-        # be identical to the output parameters because of the iterations mean
-        # that p != p0 !
-
-        # Save the fit status
-        self.fit_status = result.status
-        self.fit_success = result.success
-
-        # Save the best-fitting parameters
-        self._set_par(result.x)
-        if pe is None:
-            self.par_err = None
-        else:
-            self.par_err = np.zeros(self.np, dtype=float)
-            self.par_err[self.free] = pe
-
-        # Print the report
-        if verbose > -1:
-            self.report(fit_message=result.message)
-
-
     def report(self, fit_message=None):
         """
         Report the current parameters of the model.
@@ -1680,6 +1061,9 @@ class BisymmetricDisk:
         for i in range(ps,pe):
             print(('{0:>'+f'{max_parn_len}'+'}'+ f': {self.par[i]:.1f}').format(parn[i])
                     + (f'' if self.par_err is None else f' +/- {self.par_err[i]:.1f}'))
+        # Bisymmetry angle
+        print(('{0:>'+f'{max_parn_len}'+'}'+ f': {self.par[pe]:.1f}').format(parn[pe])
+                + (f'' if self.par_err is None else f' +/- {self.par_err[pe]:.1f}'))
         print('-'*10)
 
         print(f'First-order tangential speed parameters:')
